refactor(integration/cli): split runner into modules
